A collection of updates.
This commit includes a handful of updates to the Pokemon workshop code. First, we update to version 2.7.9 of Phaser. I've decided to include the full source version rather than the minified version. I've changed the background image to a lame hand drawn one (that we'll revisit) and have added a few extra character images. I've restructured the code a bit and have included some commented out extensions (to get the kids started).